Improving and documenting plot methods
Added some documentation, and increased flexibility of syntax, to
landlab.plot.imshow:
* Several methods can now be supplied with a nnodes-long set of values,
and the relevant subset of data will automatically be extracted and
plotted.
* Methods plotting only “active” (not closed) elements now have an
optional keyword by which the “null” plotting value can be specified.
This now defaults to the minimum element value, not 0, so the relief on
the active part of the figure is enhanced.

I left the use of “active” here, though docstrings should make clear
what this means, largely as plotting e.g., “not the closed nodes” will
probably be more common that plotting just core nodes. Could add more
methods to plot just core elements in future.