link_ca can now take tuple and int Transition IDs
Modified link_ca so transition.from_state and to_state can be *either*
ints (0,1,2,…) *or* tuples ((0,1,0), (2,3,1),… where these would be the
states 0-1 and 2/3 respectively).

Everything should still be back compatible, but this should make the
process of specifying transitions much more straightforward if you
change the total number of states!