C++ in the flow router
flow_routing/route_flow_dn (the fastscape router) now runs a bit
faster, as I’ve implemented some C++ for the loops in its helper method
flow_accum/flow_accum_bw.

This uses scipy.weave, and is set up so that in the event a compiler
isn’t found (& I think it would be in standard installs, anyway) it
will override and return to using the pure Python.

Speed gains aren’t crazy, but it definitely helped, and in part this
was an exercise testing I followed how weave works.