Adding functionality to generate_uniform_precip
Added two new methods to generate_uniform_precip,
yield_storm_interstorm_duration_intensity() and elapsed_time
(property). The former enables a generator interface so you can do
calls like

>>> for (interval, rain_rate) in yield_storm…intensity():
>>>    #the code

As if we don’t need to save the storm sequence, this will be a lot
faster. The iterator is smart enough to terminate and truncate the
final step according to the run_time.

The property method then provides a way for a driver using the
generator to get at the total time elapsed.