Adding standard component interface to Fastscape
Modifying fastscape_stream_power to follow the standard component
declarations agreed today (2/3/15).

Note that once you inherit from Component, self.grid is no longer a
valid internal name for the grid in your module, as Component makes it
a property. We may want to reconsider this choice, though changing to
self._grid is not onerous.