Wickert's gFlex
Implementing a first stab at component-izing Andy’s gFlex flexure code.
I think this way makes sense. Note we don’t hold a copy of the code;
instead, a user for now must install Andy’s module themselves (from
PyPI or github) to get the LL component to run. This avoids us having
to maintain a duplicate of his code, and removes the burden of
maintaining his code in our repository; it also emphasises the
disconnect.
However, there might well be a better way to do this, and comments
welcome.

This code hasn’t actually been checked that it works, but I think it
will. If not, only minor tweaks needed - per consultation with AW
tomorrow.

Note this follows DEJH’s current conception of an ideal LL component as
closely as possible.