gFlex handles elevs correctly
gFlex now is able to update the topographic_elevation field as well as
just lithosphere__vertical_displacement. This is not a totally
straightforward problem, as we run to steady state, and l__v_d records
the total displacement from an original zero datum, not relative to a
last tstep. However, this is now handled.