Potentiality flow router in full working order
This version of the new flow router now works very well, and is fast
and linear. Still only works on a raster for now.

Some minor issues remain with diagonal routing implemented in nodes
next to boundaries - we get a “drag” from the boundary. Easy to get
this to go away.