upgrades to overland flow component
The overland flow component has been sped up considerably and is now
stable when water begins to drain from the domain. D4 pit filling is
still an issue but JA is working on that as a separate component.

Updated the example file to make sure it cooperates with the updates to
the component.