Fixed link-fixed value node conflict
This pull resolves an issue where it was not possible to change a node
at the end of a fixed link to a fixed value node. It is now possible.

Note it is *still* impossible, by necessity and by design, to set a
node at the end of a fixed link to closed. You need to convert the
fixed link to a regular open (or closed) link first.