Perron tests
Tweak component to get it working at least rudimentarily again; this
component still has a ton of issues which need addressing (see #207).

Added very basic test to turn component over. HOWEVER, these tests
match the component as-is, not as-should-be. This means BC handling is
represented wrongly. Tests are included to prevent further component
degradation, not as true benchmarks.