More mappers
Added mappers for an max/min upwind scheme from links to nodes. This
might not be the most efficient way (some copying happens), but it’s
not too terrible.

still needed: 1. same thing, but returning a mean of all +ve/-ve links
to node. 2. same as both of previous, but returning values from a
*different* field. 3. A mapper to return which links are +ve/-ve at
each node, so these can be used as part of a more complex unwinding
scheme.