Build new helper property for FIXED_GRADIENT_BOUNDARY
Added  the property grid.fixed_gradient_boundary_node_fixed_links for
all grids. This is a helper method that returns the link ID that each
FIXED_GRADIENT_BOUNDARY node is slaved to.

The complexity in the code arises from the fact that it’s possible to
set corners of rasters as FIXED_GRADIENT_BOUNDARY, in which case they
lack a FIXED_LINK neighbour. The code resolves this by instead pointing
the corner at the (closed) link joining it to a neighbouring
FIXED_GRADIENT_BOUNDARY node (see the docstring example).