FastscapeEroder takes a threshold
The Fastscape eroded now plays nicely with thresholds, including
spatially distributed thresholds. This functionality is cythonised for
speed. Fully modernised design.

There’s a whole heap of code cleanup needed in both the class and the
supporting Cython. I *think* only two of the Cython funcs are now
needed. Bunch of redundancy in the class too.