Bug
Resolves a bug where `raster_gradients.calc_grad_at_link` would throw
an Error if any d8 flow routing had happened prior to its being called.
Issue arises because (for now) the ids of diagonal links are still
present at the end of length_of_links.

I checked the rest of the code base for similar type errors and I think
we’re now clear.