Create `BC_set_code`
Create the `BC_set_code` property for grids. It uses the clock time to
create a unique float identifier for when the boundary conditions on a
grid were last updated. The intent is that this property can be used by
components that create their own internally held BC conditions (for
speed) to test if such locally held properties need to be updated.