angle_of_link should be a property
Modify code such that `angle_of_link` is a property not a function. The
ability to specify the end of the link about which to measure angle is
replaced by a new property, `angle_of_link_about_head` (`angle_of_link`
is about the tail end, by default).

Modifications were only necessary in one place to accommodate this new
style.