active_neighbors_at_node is now a property
Had to restore some functionality to the old
`get_active_neighbors_at_node()`, as trying to move to an
all-properties usage triggered a cascade of errors in deprecated
functions still wired into flow_routing_d8 (@nicole??). I spent an hour
trying to pick it apart, but failed. The steepest descent methods still
in the grid should be removed.