Modify looped condition basic approach
Add a generic setter for looped BCs. Add basic mechanisms to update
node neighbor structures (ortho and diag). Links to come; probably
harder.

Still needs to mechanism to check of `status_at_node` and the storage
in `looped_node_properties` are internal consistent. Nodes present in
LNP but not in SaN is fine, but implies statuses have been changed and
the LNP structure needs updating. Nodes present in SaN but not LNP
implies bad setting, and an exception should be raised.

This probably stops the PerronNLDiffuser working for looped.