fix `voronoi._create_neighbors()`
Changing the layout of `_nodes` for a rectangular hex broke the logic
here. This fixes it.

This is demonstrably the only place this impacts in the grid, but it’s
possible similar effects lurk elsewhere in the components. We’ll have
to fix them as we find them, if so.