bugs in OverlandFlow timestepper; PEP8
2 minor bugs in the new time stepping routines. The second relates to
an issue already found in the LinearDiffuser, related to NaNs appearing
in the solution if a loop ever attempts to execute with no change
occurring. The giveaway symptom is the internal timestep becoming
infinite. This problem will likely recur in many different components.

Completed a PEP8 cleanup after my changes.