Make sed flux dependence explicit
This major adjustment makes the dependence on sediment in the channel
explicit - that is, sediment can now be placed in the channel, and the
component will remobilise it and work with it. To this end, a explicit
`channel_sediment__depth` field has been added. Note this is not quite
Nicole’s CHILD scheme - here we work with static depth of sediment
supplied to the component, while NMG passed the fluxes from the
hillslope fields directly to the eroder. Our approach is necessary to
fully generalise the coupling for LL, as well as hopefully being
conceptually easier to work with.

This change also addresses some deeper mass balance issues with the
component. It now demonstrably conserves mass in the active layer, and
bedrock erosion can only occur once the sed is fully mobilised.