Final stabilisation of SDE component
Main change is to enhance the pseudoimplicit part of the algorithm so
it definitely converges on a stable solution for f(qs/qc) every time
(before, we used a very quick-and-dirty technique for this).

The component now also demands the field fluvial_sediment__depth as an
output. It can also be used as an input if you want.