Changing the details of the stack to N algorithm
Initial algorithm was inspired by breadth first search algorithm but
not as efficient as a one could be.

This commit updates the algorithm to be based on existing depth first
search algorithm. This makes it faster, and maintains the ability to be
able to deal with a node providing flow to more than one boundary node
through more than one pathway (necessary for route-to-multiple methods).