virtualise sediment supply
These changes the SedFluxEroder to treat its cover layer as entirely
virtual; i.e., the topography it “sees” is a bedrock topography. (Field
names probably ought to change to reflect this, but not now?) In other
words, the `channel_sediment__depth` layer “floats above” the
topography, and changes in it don’t modify z, just dz/dt.

This change makes the component purely incisional, and thus, massively
more stable (& faster!).