Changing the default behavior of flow directors to many
Previously FlowDirectorDINF and FlowDirectorMFD created
“flow__link_to_receiver_node” and if the node flowed to itself, set the
value to -1. This is different than how the flow-director to one
methods work (value was set to node-id value).

This discrepancy was discovered by @amanaster2 while trying to use the
implicit kinematic wave component (which relies on flow director MFD)
along with the fastscape stream power component (which expects
flow-to-self nodes to have node id as the receiver not -1)

This commit fixes this issue.