Updates to EventLayers docstring
Based on a conversation I just had with @nicgaspar I've revised this docstring. 

Obviously this is the docstring of a private function, so you might ask, what;s the point. 

The intention here is that the usage of _deposit_or_erode may be confusing to someone who wants to develop with Layers. 

I've tried to improve it based on what I thought would help @ nicgaspar. Let me know if it helps. 

@mcflugen, let me know if you think anything is incorrect