Modify top edge BC
Mimicking Greg’s tweak to resolve #720, this PR resolves the “top hit
bug” in the CA when using the normal fault tectoniser as well as the
uplifter.

Note that this creates a “hard” edge at the grid top, so uplifting
nodes can never penetrate through it (as in GT’s uplifter mod).