Enhancements
1. Break off the orographic functionality into a separate func. Option
to now pass a func in that controls how this is done, rather than just
assume Singer scaling.
2. Alter how we scale the inter-event time. This version now simply
rescales the WG distribution output with a ratio, rather than modifying
the distribution itself. This is less problematic given the heavy tails
on the GEV. Modify docs to make this change clear.