Mostly comments...need to remove parcels at outlet
I went through the code and have a better idea how everything is working given the syntax. There are still some edits I don't know how to make. The code is breaking because we have to remove parcels from the system once they leave the outlet. Even if we just assign one value into a new data record and remove the others, we can add more info later. I am stuck because I don't know how to do this but maybe you do? Or else we can ask the Landlab oracle. The other thing, maybe intentional, is that the slopes are way too high and all the parcels except for those in link 5 exit the network after the first timestep.