1246 april 10
Just changed all if lat_node!=0 statements to if lat_node>0. because new way of iding boundary neighbors is with -1. so nodes that don't get lateral erosion will be either -1 or 0.  Also, I think this may have fixed the time step issue I had yesterday. I fixed it yesterday with an inelegant hack with abs(values). removed the abs(). put it back if necessary.