Added particle abrasion
Parcel abrasion has been added. This resulted in a simplifying rewrite of move_parcel_downstream. I think it is much easier to follow and understand now. I also found a few places in WC that I don't think were calculated properly -- I have updated these locations, but we should walk though this and be sure. A few other minor issues were found and are listed in the notes.