SiccarPoint/refine Taylor diffn docs (#1034)
* doctests should invoke run_one_step not other method

Modified so that run_one_step becomes a full wrapper for the `soilflux`
method, removing the need for the user to ever call it direct. This was
very confusing to me reading the docs for the first time today -
couldn’t work out what soil flux was actually doing without inspecting
the code. I think this is more (v.1!) landlabbic. I note that currently
dt takes these methods direct, and this needs to go in v.2.

* correct name in ref manual

This was inconsistent