plotting updates
With Allison's guidance, plot network links is in a nice function and plot network parcels is also in a nice function and nearly has all the desired functionality. One thing I can't figure out is how to get a contour legend for the network parcels to appear. Also, I have written a small bit of code to go from parcel attributes to link attributes and I know those bits of code exist in NST to go from link attributes to parcel attributes, but I have not put these into nice genearlizable functions yet. This would greatly improve functionality. I can work on this after AGU. This functionality will make it easier to plot colored links and parcels according to any attribute of interest. The plotting functions are now in the landlab plot folder. Calling the functions takes place at the end of NST setup. A couple to things, I found. 1) The squiggly network (which I am calling the delineated network; as opposed to the topological network of straight lines connecting nodes) has a link verticies ordered from downstream to upstream. I believe when we use actual deliniated networks, they should be upstream to downstream. For now I added a line in the locate parcel xy file to essentially flip this issue. In the future I think this should be taken out or checked for somehow. 2) For the current test network, link 4 drains to link 2 and it should drain to link 3. I am guessing this happend in the initial grid development. It doesn't matter so much as I think we would soon move on to a different "test" network. Specifically, I would like to have the "grid" attributes on a shapefile network and we pull the associated data off that shapefile in a different setup file. This is a goal for soon after AGU. Overall, two types of plots I made function well, there are just a few  small items that could use some expert insight to quickly clean up.