Paper references and minor text editing in notebooks
I believe I have fixed the issues with referencing in the paper and have added all the bibtext info for the papers referenced. Church&Ferguson2015 is in the references, but is not referenced in the paper. I am not sure if you want to cite this somewhere, Allison, or delete. I also think we could use a reference to the abrasion, either Sternberg or Parker, which reference do you prefer? Also I stepped through the tutorials and made minor text edits here and there where there were typos. I added a sentence to aid description. I also encountered one error. See my email.
Jon