calculate recurrence interval
Here is the code I wrote on calculating the recurrence interval. For whatever reason that I can not figure out, the function does not work, but the code in the function itself works. I am not sure if this should stand alone, be called external from the bed sediment initializer, or be a function within that code. I have also included an Excel data file to load, and the code in an initializeNST script I have been playing with to drive the NST. Of relevance to this function are lines 101-137 that can be incorporated somewhere else.